Add sugestões

+++ ControleDeEstoque.Application/Services/CategoriaService.cs
@@ -1,6 +1,6 @@
 ﻿using ControleDeEstoque.API.Domain.Models;
-using ControleDeEstoque.Domain.Dtos;
 using ControleDeEstoque.Domain.Interfaces;
+using ControleDeEstoque.Domain.Models.Dtos;
 using ControleDeEstoque.Domain.Shared;
 using Microsoft.AspNetCore.Mvc;
 
@@ -14,10 +14,17 @@ namespace ControleDeEstoque.Application.Services
             _genericRepository = genericRepository;
         }
 
-        [HttpGet]
-        public async Task<IEnumerable<Categoria>> ListarCategorias()
+        public async Task<IEnumerable<CategoriaDto>> ListarCategorias()
         {  
-           return await _genericRepository.GetAllAsync();
+            var listaCategoria = await _genericRepository.GetAllAsync();
+
+            var retorno = listaCategoria.Select( c => new CategoriaDto
+            {
+                Nome = c.Nome,
+                Descricao = c.Descricao
+            });
+
+            return retorno;
         }
 
         public async Task<Result> AddCategoria(CategoriaDto dados)
@@ -44,26 +51,37 @@ namespace ControleDeEstoque.Application.Services
         }
         public async Task<Result> AtualizarCategoria(CategoriaDto dados)
         {
-            var categoria = new Categoria
-            {
-                Nome = dados.Nome,
-                Descricao = dados.Descricao,
-            };
 
-            var retorno = await _genericRepository.UpdateAsync(categoria);
+            var categoriaExiste = await _genericRepository.GetAsync(p => p.Nome == dados.Nome);
 
-            if (retorno is false)
+            if (categoriaExiste == null)
+                return Result.Failure("Categoria não encontrada!");
+
+            if (dados.Nome != null)
+                categoriaExiste.Nome = dados.Nome;
+            if (dados.Descricao != null)
+                categoriaExiste.Descricao = dados.Descricao;
+
+            var retorno = await _genericRepository.SaveChangesAsync();
+
+            if (retorno is 0)
             {
                 return Result.Failure("Erro ao atualizar categoria!");
 
             }
 
-            return Result.Failure("Categoria atualizada com sucesso!");
+            return Result.Success("Categoria atualizada com sucesso!");
         }
 
         public async Task<Result> DeletarCategoria(string Nome)
         {
             var categoria = await _genericRepository.GetAsync(p => p.Nome == Nome);
+            if (categoria is null)
+            {
+                return Result.Failure("Erro ao encontrar categoria!");
+
+            }
+
 
             var retorno = await _genericRepository.DeleteAsync(categoria);
 
@@ -73,7 +91,7 @@ namespace ControleDeEstoque.Application.Services
 
             }
 
-            return Result.Failure("Categoria deletada com sucesso!");
+            return Result.Success("Categoria deletada com sucesso!");
         }
     }
 }
