+++ ControleDeEstoque.Application/Services/ProdutoService.cs
@@ -1,6 +1,6 @@
 ﻿using ControleDeEstoque.API.Domain.Models;
-using ControleDeEstoque.Domain.Dtos;
 using ControleDeEstoque.Domain.Interfaces;
+using ControleDeEstoque.Domain.Models.Dtos;
 using ControleDeEstoque.Domain.Shared;
 using Microsoft.AspNetCore.Mvc;
 
@@ -14,10 +14,21 @@ namespace ControleDeEstoque.Application.Services
             _genericRepository = genericRepository;
         }
 
-        [HttpGet]
-        public async Task<IEnumerable<Produto>> ListarProdutos()
+        public async Task<IEnumerable<ProdutoDto>> ListarProdutos()
         {  
-           return await _genericRepository.GetAllAsync();
+           var listaProdutos = await _genericRepository.GetAllAsync();
+
+            var retorno = listaProdutos.Select(p => new ProdutoDto
+            { 
+                Codigo = p.Codigo,
+                Nome = p.Nome,
+                Descricao = p.Descricao,
+                Quantidade = p.Quantidade,
+                Preco = p.Preco,
+                Categoria = p.Categoria
+            });
+
+            return retorno;
         }
 
         public async Task<Result> AddProduto(ProdutoDto dados)
@@ -50,29 +61,48 @@ namespace ControleDeEstoque.Application.Services
         }
         public async Task<Result> AtualizarProduto(ProdutoDto dados)
         {
-            var produto = new Produto
-            {
-                Codigo = dados.Codigo,
-                Nome = dados.Nome,
-                Quantidade = dados.Quantidade,
-                Preco = dados.Preco,
-                Categoria = dados.Categoria
-            };
+            var categoriaExiste = await _genericRepository.GetAsync(p => p.Nome == dados.Nome);
 
-            var retorno = await _genericRepository.UpdateAsync(produto);
+            if (categoriaExiste == null)
+                return Result.Failure("Produto não encontrado!");
 
-            if (retorno is false)
+            if (dados.Codigo != null)
+                categoriaExiste.Codigo = dados.Codigo;
+
+            if (dados.Nome != null)
+                categoriaExiste.Nome = dados.Nome;
+
+            if (dados.Descricao != null)
+                categoriaExiste.Descricao = dados.Descricao;
+
+            if (dados.Quantidade != null)
+                categoriaExiste.Quantidade = dados.Quantidade;
+
+            if (dados.Preco != null)
+                categoriaExiste.Preco = dados.Preco;
+
+            if (dados.Categoria != null)
+                categoriaExiste.Categoria = dados.Categoria;
+
+            var retorno = await _genericRepository.SaveChangesAsync();
+
+            if (retorno is 0)
             {
                 return Result.Failure("Erro ao atualizar produto!");
 
             }
 
-            return Result.Failure("Produto atualizado com sucesso!");
+            return Result.Success("Produto atualizado com sucesso!");
         }
 
         public async Task<Result> DeletarProduto(string Codigo)
         {
             var produto = await _genericRepository.GetAsync(p => p.Codigo == Codigo);
+
+            if (produto is null)
+            {
+                return Result.Failure("Produto não encontrado!");
+            }
 
             var retorno = await _genericRepository.DeleteAsync(produto);
